Don't clobber docker repo in test_DockerImage, keep it for test_Knot
